SKEU-3: 페이지네이션 페이지 및 사이즈 타입 변경

+++ src/main/java/com/everyonewaiter/global/support/Paging.java
@@ -12,16 +12,16 @@ public class Paging<T> {
   private final List<T> content = new ArrayList<>();
 
   @Schema(description = "현재 페이지 번호", example = "1")
-  private final int page;
+  private final long page;
 
   @Schema(description = "조회 데이터 수", example = "20")
-  private final int size;
+  private final long size;
 
   @Schema(description = "FastForward 페이지 사이즈", example = "5")
   private final int pageSkipSize;
 
   @Schema(description = "FastForward 계산을 위해 조회된 데이터 수", example = "101")
-  private final int count;
+  private final long count;
 
   @Schema(description = "다음 페이지 여부", example = "true")
   private final boolean hasNext;
@@ -36,12 +36,12 @@ public class Paging<T> {
   private final boolean isLast;
 
   @Schema(description = "FastForward 페이지 번호", example = "6")
-  private final int fastForwardPage;
+  private final long fastForwardPage;
 
   @Schema(description = "FastBackward 페이지 번호", example = "1")
-  private final int fastBackwardPage;
+  private final long fastBackwardPage;
 
-  public Paging(List<T> content, int count, PagingRequest pagingRequest) {
+  public Paging(List<T> content, long count, PagingRequest pagingRequest) {
     this.content.addAll(content);
     this.page = pagingRequest.page();
     this.size = pagingRequest.size();
@@ -51,7 +51,7 @@ public class Paging<T> {
     this.hasPrevious = page > 1;
     this.isFirst = page == 1;
     this.isLast = !hasNext;
-    this.fastForwardPage = Math.min(page + pageSkipSize, (int) Math.ceil((double) count / size));
+    this.fastForwardPage = Math.min(page + pageSkipSize, (long) Math.ceil((double) count / size));
     this.fastBackwardPage = Math.max(page - pageSkipSize, 1);
   }
